Move now testing to test module
